[CP-3674] ManageFiles domain restructured and naming unified
