+++ libs/devices/common/ui/src/index.ts
@@ -12,7 +12,7 @@ export * from "./lib/devices-indicator/devices-indicator"
 export * from "./lib/switching-loader/device-switching-loader"
 export * from "./lib/device-troubleshooting/device-troubleshooting"
 export * from "./lib/welcome-screen/welcome-screen"
-export * from "./lib/views/manage-files/manage.files"
+export * from "./lib/views/manage-files/manage-files"
 export * from "./lib/views/manage-files/manage-files.types"
 export * from "./lib/views/overview/overview"
 export * from "./lib/views/overview/overview-status-item"
