+++ libs/devices/common/ui/src/lib/views/manage-files/manage-files-content/manage-files-other-files.tsx
@@ -7,7 +7,7 @@ import { FunctionComponent } from "react"
 import styled from "styled-components"
 import { IconType } from "app-theme/models"
 import { formatBytes, Icon, Tooltip, Typography } from "app-theme/ui"
-import { manageFilesMessages } from "./manage-files.messages"
+import { manageFilesMessages } from "../manage-files.messages"
 
 export interface ManageFilesOtherFilesProps {
   otherFiles: { name: string }[]
