+++ libs/devices/common/ui/src/lib/views/manage-files/manage-files-content/manage-files-table-section.tsx
@@ -3,7 +3,7 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-import { FileManagerFileMap } from "./manage-files.types"
+import { FileManagerFileMap } from "../manage-files.types"
 
 export interface ManageFilesTableSectionProps {
   fileMap: FileManagerFileMap
