+++ libs/devices/common/ui/src/lib/views/manage-files/manage-files-delete-flow/manage-files-delete-confirm-modal.tsx
@@ -22,7 +22,7 @@ export interface ManageFilesConfirmDeleteModalProps {
   }
 }
 
-export const ManageFilesConfirmDeleteModal: FunctionComponent<
+export const ManageFilesDeleteConfirmModal: FunctionComponent<
   ManageFilesConfirmDeleteModalProps
 > = ({ opened, onCancel, onConfirm, onClose, fileCount, messages }) => {
   return (
