+++ libs/devices/common/ui/src/lib/views/manage-files/manage-files-delete-flow/manage-files-delete-failed-modal.tsx
@@ -5,8 +5,8 @@
 
 import { FunctionComponent, useMemo } from "react"
 import { formatMessage, Messages } from "app-localize/utils"
-import { FileManagerFile } from "./manage-files.types"
-import { ManageFilesFailedModal } from "./manage-files-failed-modal"
+import { FileManagerFile } from "../manage-files.types"
+import { ManageFilesGenericFailedModal } from "../manage-files-generic-failed-modal"
 
 export interface ManageFilesDeleteFailedModalProps {
   opened: boolean
@@ -54,7 +54,7 @@ export const ManageFilesDeleteFailedModal: FunctionComponent<
   const failedFilesList = total > 1 && !isAllFailed ? failedFiles : undefined
 
   return (
-    <ManageFilesFailedModal
+    <ManageFilesGenericFailedModal
       opened={opened}
       onClose={onClose}
       title={title}
