+++ libs/devices/common/ui/src/lib/views/manage-files/manage-files-transfer-flow/manage-files-transfer-progress-modal.tsx
@@ -20,7 +20,7 @@ export interface ManageFilesTransferringModalProps {
   progressBarMessage: string
 }
 
-export const ManageFilesTransferringModal: FunctionComponent<
+export const ManageFilesTransferProgressModal: FunctionComponent<
   ManageFilesTransferringModalProps
 > = ({
   opened,
