+++ libs/devices/common/ui/src/lib/views/manage-files/manage-files-transfer-flow/use-manage-files-validate.ts
@@ -4,8 +4,8 @@
  */
 
 import { useCallback, useMemo } from "react"
+import { FileManagerFile } from "../manage-files.types"
 import { validateSelectedFiles } from "./validate-selected-files"
-import { FileManagerFile } from "./manage-files.types"
 
 interface UseManageFilesValidateArgs {
   fileMap: Record<string, FileManagerFile>
