+++ libs/devices/common/ui/src/lib/views/manage-files/manage-files-transfer-flow/validate-selected-files.ts
@@ -11,7 +11,7 @@ import {
   TransferErrorName,
   ValidationSummary,
   ValidationSummaryType,
-} from "./manage-files.types"
+} from "../manage-files.types"
 
 export const calculateAndFormatAvailableSpace = (
   freeSpaceBytes: number,
